* extend search radius for gray-white gradient correction * fix weights for self-intersections * normalize oversampling weight by local area for find_image_boundaries   for improved convergence rate on non-uniform meshes * extension of gray-white gradient for histology (BigBrain) * support for fitting of white surface using chamfer map for Laplacian field * improved mesh adaptation (convergence, swapping, shrinkage-free   Taubin smoothing) * add surf_surf_interpolate.c and surface_angles.c

+++ src/sphere.cc
@@ -374,7 +374,7 @@ int make_sphere( int sizes[MAX_DIMENSIONS], short * val ) {
 	    }
 	  }
           if( count ) {
-            levels2[ii] = rint( (float)sum / count );
+            levels2[ii] = (short)rint( (float)sum / count );
           }   
         }
       }
